Implement archive extension
Mithrandir now allows user to archive tasks according to a given
index. The archived task is removed from the current task list and
is stored in a file named Archive.txt.

Fix bug of unloaded task list

The saved task list is not loaded from Save.txt when GUi launches.
Extract out loadTaskList as a public delegate function in
Application.java such that GUi can call this function when
instantiating.